fix: SecretNFT preview adjusted
